Adds a csv file upload button to the dashboard
Allows users to upload a csv file via a button on the dashboard, this
file is then saved in the backend.

This is a combination of 7 commits.

This is a list of commit messages:

Initial commit of laravel project

implement eslint and prettier for code quality

lint

lint

Add file upload button to dashboard

remove vite temp files

gitignore vite